Refactor Hash functions and hash tables; add topmost embedding for multi-width multiplication
Hash table infrastructure:
- Remove HASHBASE/HASHSETBASE constants; tables now do % numBuckets internally
- Change all Hash() signatures from unsigned int Hash(unsigned long modsize) to size_t Hash()
- Add PTR_ALIGN_SHIFT constexpr for correct 64-bit pointer hashing
- Set bucket sizes to 2M (HASH_NUM_BUCKETS, HASHSET_NUM_BUCKETS)

Topmost embedding for CFLOBDD multiplication:
- Decouple CFLOBDDMaxLevel from bit-width via virtualMaxLevel/bottomLevel/stride
- Add per-width build targets (cflobdd32..cflobdd4096) with NUM_BITS macro
- Restructure tests with command-line dispatch (shiftadd-all, karatsuba, crt, etc.)
- Add --verbose flag support and MakeAssignment helper

Makefile: track WCFLOBDD_SUPPORTED-dependent objects to prevent stale .o linker errors

+++ CFLOBDD/hash.h
@@ -13,8 +13,8 @@
 // lookup and insert operations.
 //
 // NOTE: The key type (KeyT) must provide:
-//       1. a Hash member function that takes one parameter k and
-//          returns an integer in the range 0..k-1
+//       1. a Hash member function that returns a size_t hash value
+//          (the table performs % numBuckets internally)
 //       2. a definition of std::ostream &operator<<(std::ostream, const KeyT &)
 //       3. a definition of bool operator==(const KeyT &)
 //       The item type (ItemT) must also provide operator<<.
@@ -80,7 +80,7 @@ template <class KeyT, class ItemT> class Hashtable
         
     // other operations
 
-    int Size() const;
+    unsigned long Size() const;
     bool Lookup(KeyT k) const;
     bool Fetch(KeyT k, ItemT &i) const;
 	//ItemT & operator [] (KeyT k);
@@ -89,7 +89,7 @@ template <class KeyT, class ItemT> class Hashtable
 
   private:
     const int numBuckets;
-    int mySize;                        // current size
+    unsigned long mySize;                // current size
     apvector<List<Pair<KeyT, ItemT> *> > *myItems; // pointer to array of items
 };
 
@@ -110,7 +110,7 @@ template <class KeyT, class ItemT> class HashtableIterator
 
   private:
     apvector<List<Pair<KeyT, ItemT> *> > *myItems; // pointer to array of items
-    int mySize;                        // current size
+    unsigned long mySize;                // current size
     int numBuckets;
     int k;
     int numAccessed;
